Time: 38 ms (91.12%), Space: 18.9 MB (38.24%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -14,7 +14,7 @@ public:
     long high=maxEatable;
     int ans=-1;
     while(low<=high){
-        long long mid=(low+high)/2;
+         long mid=(low+high)/2;
         long totalH=func(v,mid);
         if(totalH<=h){
             ans=mid;
